ctl instance stop-all test
